requirements.txt
Flask
python-dotenv
pandas
google-generativeai
